Break up components/auth (#15713)
